feat: Add initial page structures, sections, and data for home, events, and organizations
