+++ client-side-ts/src/data/index.ts
@@ -1,0 +1,6 @@
+export * from "./home-data";
+export * from "./about-data";
+export * from "./events-data";
+export * from "./organization-data";
+export * from "../@types/sections";
+
