+++ client-side-ts/src/pages/events/Events.tsx
@@ -1,0 +1,13 @@
+import { EventsBanner } from './sections/EventsBanner';
+import { UpcomingEventsSection } from './sections/UpcomingEventsSection';
+import { PastEvents } from './sections/PastEvents';
+
+export const Events = () => {
+  return (
+    <>
+      <EventsBanner />
+      <UpcomingEventsSection />
+      <PastEvents />
+    </>
+  );
+};
