+++ client-side-ts/src/pages/events/sections/EventsBanner.tsx
