+++ client-side-ts/src/pages/home/sections/GetInvolvedCard.tsx
@@ -1,23 +1,22 @@
 import { Card, CardDescription, CardTitle } from '@/components/ui/card';
 import { cn } from '@/lib/utils';
-import { ArrowRight } from 'lucide-react';
 import type { LucideIcon } from 'lucide-react';
 import announcement from '@/assets/announcement.png';
 
 interface GetInvolvedCardProps {
-  title: string;
-  description: string;
-  icon: LucideIcon;
-  image?: string;
-  className?: string;
+    title: string;
+    description: string;
+    icon: LucideIcon;
+    image?: string;
+    className?: string;
 }
 
 export const GetInvolvedCard = ({
-  title,
-  description,
-  icon: Icon,
-  image,
-  className,
+    title,
+    description,
+    icon: Icon,
+    image,
+    className,
 }: GetInvolvedCardProps) => (
     <Card className={cn(
         "group relative overflow-hidden border-none shadow-lg p-0 transition-all duration-500 hover:-translate-y-1 md:hover:-translate-y-2 flex flex-col min-h-[280px] sm:min-h-[320px] md:min-h-[360px] lg:min-h-[400px] rounded-2xl sm:rounded-3xl w-full",
@@ -50,13 +49,13 @@ export const GetInvolvedCard = ({
         {/* Description & CTA - Absolute Bottom */}
         <div className="absolute bottom-0 left-0 right-0 z-10 p-5 sm:p-6 md:p-8">
             <div className="space-y-2 sm:space-y-3">
-               <CardTitle className="text-xl sm:text-2xl md:text-3xl font-bold tracking-tight text-white group-hover:text-primary transition-colors duration-300 drop-shadow-lg">
-                {title}
-            </CardTitle>
+                <CardTitle className="text-xl sm:text-2xl md:text-3xl font-bold tracking-tight text-white group-hover:text-primary transition-colors duration-300 drop-shadow-lg">
+                    {title}
+                </CardTitle>
                 <CardDescription className="text-gray-300/90 leading-relaxed text-xs sm:text-sm md:text-base max-w-[90%]">
                     {description}
                 </CardDescription>
-                
+
                 {/* CTA Button */}
                 {/* <div className="pt-2 md:pt-3 flex">
                     <button className="inline-flex items-center gap-2 text-xs md:text-sm font-semibold text-white/90 group-hover:text-primary transition-all duration-300">
